Create Tensorflow_README
This model is a binary classifier, meaning that it is designed to predict one of two possible outcomes for a given input. In this case, the model is likely used to predict whether a customer will churn (cancel their subscription) or not based on their data.

The model takes as input a vector of features representing the customer, such as their age, gender, location, and usage patterns. It then uses these features to calculate a probability that the customer will churn. If the probability is greater than 0.5, the model predicts that the customer will churn. Otherwise, it predicts that the customer will not churn.

The model was trained on a dataset of customer data, and its performance was evaluated using the accuracy metric. The accuracy metric measures the proportion of correctly classified examples. The model achieved an accuracy score of 83.33% on the test data, which means that it correctly predicted the churn status of 83.33% of the customers in the test set.

The model can be used to identify customers who are at risk of churning, so that businesses can take steps to prevent them from canceling their subscriptions. For example, a business might offer a discount or a special promotion to a customer who is predicted to churn.